requirements.txt
pandas
numpy
python-dotenv
certifi
pymongo[srv]==3.6
pymongo
scikit-learn
#-e .  